Split CLI to utils file
