+++ raster-cli/src/main.cpp
